C++ compatibility changes
These changes will allow the component to be used in C++ Builder. I have tested the changes with RAD Studio 10.1 Berlin (using a JavaScript file), and whilst the syntax highlighting works there is a strange bug where clicking inside the editor selects all lines from the bottom of the editor up to and including the current line. I am not familiar enough with the component yet to work out why this is happening.